Refactor blob storage
