+++ app/config/storage/index.js
@@ -1,0 +1,5 @@
+const blobConfig = require('./blob')
+
+module.exports = {
+  blobConfig
+}
