+++ app/storage/get-blob-client.js
@@ -1,0 +1,18 @@
+const { BlobServiceClient } = require('@azure/storage-blob')
+const { DefaultAzureCredential } = require('@azure/identity')
+const { blobConfig } = require('../config/storage')
+
+let blobServiceClient
+
+if (blobConfig.useConnectionString) {
+  console.log('Using connection string for BlobServiceClient')
+  blobServiceClient = BlobServiceClient.fromConnectionString(blobConfig.connectionString)
+} else {
+  console.log('Using DefaultAzureCredential for BlobServiceClient')
+  const uri = `https://${blobConfig.storageAccount}.blob.core.windows.net`
+  blobServiceClient = new BlobServiceClient(uri, new DefaultAzureCredential())
+}
+
+module.exports = {
+  blobServiceClient
+}
